+++ pkg/tenant/auth.go
@@ -10,7 +10,7 @@ import (
 	middleware "github.com/rotationalio/ensign/pkg/quarterdeck/middleware"
 	"github.com/rotationalio/ensign/pkg/tenant/api/v1"
 	"github.com/rotationalio/ensign/pkg/tenant/db"
-	"github.com/rotationalio/ensign/pkg/utils/emails"
+	"github.com/rotationalio/ensign/pkg/utils/sendgrid"
 	"github.com/rotationalio/ensign/pkg/utils/ulid"
 	"github.com/rs/zerolog/log"
 )
@@ -82,7 +82,7 @@ func (s *Server) Register(c *gin.Context) {
 	if s.conf.SendGrid.Enabled() {
 		go func() {
 			name := strings.Split(params.Name, "")
-			contact := &emails.Contact{
+			contact := &sendgrid.Contact{
 				Email:     params.Email,
 				FirstName: name[0],
 			}
