+++ pkg/utils/emails/contacts.go
@@ -7,44 +7,20 @@ import (
 	"fmt"
 	"net/http"
 
+	sg "github.com/rotationalio/ensign/pkg/utils/sendgrid"
 	"github.com/sendgrid/sendgrid-go"
 )
 
-const (
-	sgHost     = "https://api.sendgrid.com"
-	sgContacts = "/v3/marketing/contacts"
-)
-
-type AddContact struct {
-	ListIDs  []string   `json:"list_ids"`
-	Contacts []*Contact `json:"contacts"`
-}
-
-type Contact struct {
-	FirstName    string        `json:"first_name"`
-	LastName     string        `json:"last_name"`
-	Email        string        `json:"email"`
-	Country      string        `json:"country"`
-	CustomFields *CustomFields `json:"custom_fields"`
-}
-
-// TODO: make custom fields request to get field IDs rather than hardcoding.
-type CustomFields struct {
-	Title                string `json:"e1_T"`
-	Organization         string `json:"e2_T"`
-	CloudServiceProvider string `json:"e3_T"`
-}
-
 // AddContact adds a contact to the SendGrid marketing contacts list.
-func (m *EmailManager) AddContact(contact *Contact) error {
+func (m *EmailManager) AddContact(contact *sg.Contact) error {
 	if !m.conf.Enabled() {
 		return errors.New("sendgrid is not enabled, cannot add contact")
 	}
 
 	// Create the SendGrid request
 	var buf bytes.Buffer
-	sgdata := &AddContact{
-		Contacts: []*Contact{contact},
+	sgdata := &sg.AddContact{
+		Contacts: []*sg.Contact{contact},
 	}
 
 	if m.conf.EnsignListID != "" {
@@ -56,7 +32,7 @@ func (m *EmailManager) AddContact(contact *Contact) error {
 	}
 
 	// Execute the SendGrid request
-	req := sendgrid.GetRequest(m.conf.APIKey, sgContacts, sgHost)
+	req := sendgrid.GetRequest(m.conf.APIKey, sg.Contacts, sg.Host)
 	req.Method = http.MethodPut
 	req.Body = buf.Bytes()
